requirements.txt
Flask
python-dotenv
flask_pymongo
flask_restful
pymongo[srv]